Fix MSVC warning C4267
QOpenGLBuffer::allocate expects an int, not a size_t:

  warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data

+++ src/charts/glwidget.cpp
@@ -355,7 +355,7 @@ void GLWidget::render(bool selection)
             }
             vbo->bind();
             if (dirty) {
-                vbo->allocate(data->array.constData(), data->array.count() * sizeof(GLfloat));
+                vbo->allocate(data->array.constData(), int(data->array.count() * sizeof(GLfloat)));
                 dirty = false;
                 m_selectionRenderNeeded = true;
             }
